feat. board 상세 보기 페이지 추가
