+++ src/App.jsx
@@ -8,8 +8,8 @@ import StartPage from './pages/StartPage';
 import NotFoundPage from './pages/NotFoundPage';
 import MainPage from './pages/MainPage';
 import LoginPage from './pages/LoginPage';
-import BoardDetailPage from './pages/BoardDetailPage';
-import WritePage from './pages/board/WritePage';
+import BoardDetailPage from './pages/board/BoardDetailPage';
+import BoardWritePage from './pages/board/BoardWritePage';
 
 // 전역 스타일
 const globalStyle = css`
@@ -44,8 +44,8 @@ const App = () => {
 
             {/* url parameter 활용하여, 자유게시판목록 디테일 들어감 */}
             <Route path="/board" element={<BoardPage />} />
-            <Route path="/board/view/:postId" element={<BoardDetailPage />} />{' '}
-            <Route path="/board/write" element={<WritePage />} />
+            <Route path="/board/:postId" element={<BoardDetailPage />} />{' '}
+            <Route path="/board/write" element={<BoardWritePage />} />
           </Route>
 
           {/* Layout 미적용 페이지 */}
