+++ src/components/board/MarkdownViewer.jsx
@@ -1,4 +1,5 @@
-import { css } from '@emotion/react';
+/** @jsxImportSource @emotion/react */
+import { css } from "@emotion/react";
 import Markdown from "react-markdown";
 import rehypeSanitize, { defaultSchema } from "rehype-sanitize";
