+++ src/pages/board/BoardWritePage.jsx
@@ -1,7 +1,7 @@
 import { css } from "@emotion/react";
 import MarkdownEditor from "../../components/board/MarkdownEditor"
 
-export default function WritePage() {
+export default function BoardWritePage() {
   return (
     <div>
       <MarkdownEditor />
